SIH_adaptive_learning/requirements.txt
asgiref==3.7.2
Django==4.2.8
<<<<<<< HEAD
django-filter==23.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
=======
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
>>>>>>> 335dd2464cdd8e13d40babf5f3fcbb8a822c1c2a
